Commands execute as designated times, plus new commands
A timestamp for the command is now read from each line in the text file. Assuming the commands are in order within the timeline, they will execute when their designated time is reached. Added translate and animate commands as well. Translate moves the object by the specified offset, and animate applies the specified animator to the object.