Implemented LOOKAT command and general view improvement
LOOKAT <gameObject name> will set the following camera to the specified object at the given time. Changed following view based on object bounds to view the object in a consistent way.  This also allows for zooming in and out on the object via scroll wheel. Also tweaked path creation to start at 0,0,0 to make setting the path start more predictable.