Altered timescale handling
Removed TIMESCALE command,  choosing to use CALL <func> <timeScale> instead. Each is quite similar in use, with only slight differences, and the CALL method is preferable for readability and nested reference behavior. A called function will have its parent's scale by default, or 1 in the case of the main function. The scale will multiply if specified. The playback time  scale is now independent from the in text timescale, making the interaction cleaner. They will also compound in the playback as expected.